+++ internal/config/config.go
@@ -158,10 +158,6 @@ const (
 type Options struct {
 	Mode              Mode
 	RemoveWhitespace  bool
-	RemoveConsole     bool
-	RemoveDebugger    bool
-	DebugToolName     string
-	RemoveDebugTool   bool
 	MinifyIdentifiers bool
 	MangleSyntax      bool
 	CodeSplitting     bool
@@ -222,6 +218,11 @@ type Options struct {
 
 	SourceMap SourceMap
 	Stdin     *StdinInfo
+
+	RemoveConsole   bool
+	RemoveDebugger  bool
+	DebugTool       string
+	RemoveDebugTool bool
 }
 
 type InjectedFile struct {