+++ lib/common.ts
@@ -108,6 +108,8 @@ function pushCommonFlags(flags: string[], options: CommonOptions, keys: OptionKe
   let footer = getFlag(options, keys, 'footer', mustBeString);
   let removeConsole = getFlag(options, keys, 'removeConsole', mustBeBoolean);
   let removeDebugger = getFlag(options, keys, 'removeDebugger', mustBeBoolean);
+  let debugTool = getFlag(options, keys, 'debugTool', mustBeString);
+  let removeDebugTool = getFlag(options, keys, 'removeDebugTool', mustBeBoolean);
 
   if (target) {
     if (Array.isArray(target)) flags.push(`--target=${Array.from(target).map(validateTarget).join(',')}`)
@@ -140,6 +142,8 @@ function pushCommonFlags(flags: string[], options: CommonOptions, keys: OptionKe
 
   if (removeConsole) flags.push('--remove-console');
   if (removeDebugger) flags.push('--remove-debugger');
+  if (debugTool) flags.push(`--debug-tool=${debugTool}`);
+  if (removeDebugTool) flags.push('--remove-debug-tool');
 }
 
 function flagsForBuildOptions(callName: string, options: types.BuildOptions, isTTY: boolean, logLevelDefault: types.LogLevel, writeDefault: boolean):